% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_na_imputations.R
\name{ggplot_na_imputations}
\alias{ggplot_na_imputations}
\title{Visualize Imputed Values}
\usage{
ggplot_na_imputations(
  x_with_na,
  x_with_imputations,
  x_with_truth = NULL,
  x_axis_labels = NULL,
  title = "Imputed Values",
  subtitle = "Visualization of missing value replacements",
  xlab = "Time",
  ylab = "Value",
  color_points = "steelblue",
  color_imputations = "indianred",
  color_truth = "seagreen3",
  color_lines = "lightslategray",
  shape_points = 16,
  shape_imputations = 18,
  shape_truth = 16,
  size_points = 1.5,
  size_imputations = 2.5,
  size_truth = 1.5,
  size_lines = 0.5,
  linetype = "solid",
  connect_na = T,
  legend = TRUE,
  legend_size = 5,
  label_known = "known values",
  label_imputations = "imputed values",
  label_truth = "ground truth",
  theme = ggplot2::theme_linedraw()
)
}
\arguments{
\item{x_with_na}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs before imputation. This parameter and
x_with_imputations (and eventually x_with_truth if available) have to be set. The rest of the parameters are only needed if you want to adjust the default design.}

\item{x_with_imputations}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs replaced by imputed values. This parameter and
x_with_na (and eventually x_with_truth if available) have to be set. The rest of the parameters are only needed if you want to adjust the default design.}

\item{x_with_truth}{Numeric Vector or Time Series (\code{\link{ts}}) object with the real values (can be set to NULL if not known).}

\item{x_axis_labels}{For adding specific x-axis labels. Takes a vector (with the same length as x)
of either Date or POSIXct objects as an input. Default (NULL) is using the
observation number as  x-axis tick labels.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{xlab}{Label for x axis.}

\item{ylab}{Label for y axis.}

\item{color_points}{Color of the symobls for the normal non-NA observations.}

\item{color_imputations}{Color of the symbols for the imputed values.}

\item{color_truth}{Color of the symbols for the ground truth of the NA values.}

\item{color_lines}{Color for the lines connecting the observations/points.}

\item{shape_points}{Shape of the symobls for the normal non-NA observations.}

\item{shape_imputations}{Shape of the symbols for the imputed values.}

\item{shape_truth}{Shape of the symbols for the ground truth of the NA values.}

\item{size_points}{Size of the symobls for the normal non-NA observations.}

\item{size_imputations}{Size of the symbols for the imputed values.}

\item{size_truth}{Size of the symbols for the ground truth of the NA values.}

\item{size_lines}{Size for the lines connecting the observations/points.}

\item{linetype}{Linetype for the lines connecting the observations/points.}

\item{connect_na}{If TRUE the imputations/ground truth values are connected to the normal
non-NA observations in the plot. Otherwise there are no connecting lines between symbols in NA areas.}

\item{legend}{If TRUE a legend is added at the bottom.}

\item{legend_size}{Size of the symbols used in the legend.}

\item{label_known}{Legend label for the normal non-NA observations.}

\item{label_imputations}{Legend label for the imputed values.}

\item{label_truth}{Legend label for the ground truth of the NA values.}

\item{theme}{Set a theme for ggplot2. Default is \code{\link[ggplot2]{theme_linedraw}}}
}
\description{
Visualize the imputed values in a time series.
}
\details{
This plot can be used, to visualize imputed values for a time series.
Imputed values (filled NA gaps) are shown in a different color than the other values.
If real values (truth) for the NA gaps are known, they are added in a different color.
}
\examples{
# Example 1: Visualize the values that were imputed by na.mean in the time series
impMean.Airgap <- na_mean(tsAirgap)
ggplot_na_imputations(tsAirgap, impMean.Airgap)


# Example 2: Visualize the imputed values by na_locf and the true values in the time series
impLOCF.Airgap <- na_locf(tsAirgap)
ggplot_na_imputations(tsAirgap, impLOCF.Airgap, tsAirgapComplete)

# Example 3: Same as example 1, just written with pipe operator
tsAirgap \%>\%
  na_mean() \%>\%
  ggplot_na_imputations(x_with_na = tsAirgap)
}
\seealso{
\code{\link[imputeTS]{ggplot_na_distribution}},\code{\link[imputeTS]{ggplot_na_intervals}},
\code{\link[imputeTS]{ggplot_na_gapsize}}
}
\author{
Steffen Moritz
}
