% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.forecast.R
\name{na.forecast}
\alias{na.forecast}
\title{Missing Value Imputation by Forecast/Backcast combinations}
\usage{
na.forecast(x, weighting = "linear")
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}}) object in which missing values shall be replaced}

\item{weighting}{Weighting between the forecaste and the backcast. Accepts the following input:
\itemize{
   \item{"equal" - Forecast and Backcast are weighted equally. The mean of both values for an NA position is used as final result.}
   \item{"quantity" - Forecast and Backcast are weighted according to the quantity of data available for model building.(see Details section) }
   }}

\item{...}{Additional parameters to be passed through to \link{approx} or \link{spline}interpolation functions}
}
\value{
Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}}) object (dependent on given input at parameter x)
}
\description{
Uses a combination of time series forecasts / backcasts for imputation. With multiple different forecasting methods usable.
}
\details{
Missing values get replaced by values 

The implementation uses the forecast package
}
\examples{
#Create Time series with missing values
x <- ts(c(2,3,4,5,6,NA,7,8))

#Perform linear interpolation
na.interpolation(x)

#Perform spline interpolation
na.interpolation(x, option ="spline")

}
\author{
Steffen Moritz
}
\seealso{
\code{\link[imputeTS]{na.interpolation}},
\code{\link[imputeTS]{na.kalman}}, \code{\link[imputeTS]{na.locf}},
 \code{\link[imputeTS]{na.ma}}, \code{\link[imputeTS]{na.mean}},
 \code{\link[imputeTS]{na.random}}, \code{\link[imputeTS]{na.replace}},
 \code{\link[imputeTS]{na.seadec}}, \code{\link[imputeTS]{na.seasplit}}
}

