% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNA_distribution.R
\name{plotNA_distribution}
\alias{plotNA_distribution}
\title{Visualize Distribution of Missing Values}
\usage{
plotNA_distribution(x, colPoints = "steelblue",
  colBackgroundMV = "indianred2", main = "Distribution of NAs",
  xlab = "Time", ylab = "Value", pch = 20, cexPoints = 2.5,
  col = "black", theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}}) object containing NAs}

\item{colPoints}{Point color for observations}

\item{colBackgroundMV}{Background color for NA sequences}

\item{main}{Main title}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{pch}{Plotting 'character', i.e., symbol to use}

\item{cexPoints}{Character (or symbol) expansion: a numerical vector}

\item{col}{Line color}

\item{theme}{Set a theme for ggplot2. Default is \code{\link[ggplot2]{theme_minimal}}}

\item{...}{These parameters are passed to \code{\link[ggplot2]{geom_line}}}
}
\description{
Visualize the distribution of missing values within a time series.
}
\details{
This function visualizes the distribution of missing values within a time series.
If a value is NA, the background is colored differently. This gives a good overview
of where most missing values occur.
}
\examples{
# Example 1: Visualize the missing values in x
# x <- ts(c(1:11, 4:9,NA,NA,NA,11:15,7:15,15:6,NA,NA,2:5,3:7))
# plotNA_distribution(x)

# Example 2: Visualize the missing values in tsAirgap time series
plotNA_distribution(tsAirgap)

# Example 3: Same as example 1, just written with pipe operator
x <- ts(c(1:11, 4:9, NA, NA, NA, 11:15, 7:15, 15:6, NA, NA, 2:5, 3:7))
x \%>\% plotNA_distribution()
}
\seealso{
\code{\link[imputeTS]{plotNA_distributionBar}},
 \code{\link[imputeTS]{plotNA_gapsize}}, \code{\link[imputeTS]{plotNA_imputations}}
}
\author{
Steffen Moritz
}
