% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNA_imputations.R
\name{plotNA_imputations}
\alias{plotNA_imputations}
\title{Visualize Imputed Values}
\usage{
plotNA_imputations(
  x_with_na,
  x_with_imputations,
  x_with_truth = NULL,
  xlab = "Time",
  ylab = "Value",
  legend = TRUE,
  main = "Visualization Imputed Values",
  color_lines = "black",
  color_points = "steelblue2",
  color_imputations = "indianred2",
  color_truth = "green3",
  shape_points = 16,
  size_points = 2,
  size_lines = 0.5,
  linetype = "solid",
  theme = ggplot2::theme_minimal(),
  ...
)
}
\arguments{
\item{x_with_na}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs before imputation}

\item{x_with_imputations}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs replaced by imputed values}

\item{x_with_truth}{Numeric Vector or Time Series (\code{\link{ts}}) object with the real values. (can be set to NULL if not known)}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{legend}{If TRUE a legend is added at the bottom}

\item{main}{Main title}

\item{color_lines}{Color of lines connecting non-NA observations}

\item{color_points}{Color of normal non-NA observations}

\item{color_imputations}{Color of imputed values}

\item{color_truth}{Color of real values (truth) for the NA values}

\item{shape_points}{Either an integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{size_points}{A numerical value giving the size of points.}

\item{size_lines}{A numerical value giving the size of lines.}

\item{linetype}{Linetype used for connecting the points}

\item{theme}{Set a theme for ggplot2. Default is \code{\link[ggplot2]{theme_minimal}}}

\item{...}{Additional graphical parameters can be passed to \code{\link[ggplot2]{geom_line}}}
}
\description{
Visualize the imputed values in a time series.
}
\details{
This plot can be used, to visualize imputed values for a time series.
Imputed values (filled NA gaps) are shown in a different color than the other values.
If real values (truth) for the NA gaps are known, they are added in a different color.
}
\examples{
# Example 1: Visualize the values that were imputed by na.mean in the time series
impMean.Airgap <- na_mean(tsAirgap)
plotNA_imputations(tsAirgap, impMean.Airgap)


# Example 2: Visualize the imputed values by na_locf and the true values in the time series
impLOCF.Airgap <- na_locf(tsAirgap)
plotNA_imputations(tsAirgap, impLOCF.Airgap, tsAirgapComplete)

# Example 3: Same as example 1, just written with pipe operator
tsAirgap \%>\%
  na_mean() \%>\%
  plotNA_imputations(x_with_na = tsAirgap)
}
\seealso{
\code{\link[imputeTS]{plotNA_distribution}},\code{\link[imputeTS]{plotNA_distributionBar}},
 \code{\link[imputeTS]{plotNA_gapsize}}
}
\author{
Steffen Moritz
}
