% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int.precheck.R
\name{precheck}
\alias{precheck}
\title{Pre-Check Imputation Input (Internal function)}
\usage{
precheck(data, na.identifier)
}
\arguments{
\item{data}{Supposed to be a univariate time series}

\item{na.identifier}{Supposed to be a character string or numeric value.
If another value than NA indicates missing values this is specified with this parameter.}
}
\value{
Time Series (\code{\link{ts}}) object that fulfills the requirements
}
\description{
Internal function for pre-checking the input of imputation functions
}
\details{
Checks, if data is really a univariate time series and checks that na.identifier is no list or vector.
}
\author{
Steffen Moritz
}

