% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_na_distribution.R
\name{ggplot_na_distribution}
\alias{ggplot_na_distribution}
\title{Visualize Distribution of Missing Values}
\usage{
ggplot_na_distribution(
  x,
  x_axis_labels = NULL,
  color_points = "steelblue",
  color_lines = "steelblue2",
  color_missing = "indianred",
  color_missing_border = "indianred",
  alpha_missing = 0.5,
  title = "Distribution of Missing Values",
  subtitle = "Time Series with highlighted missing regions",
  xlab = "Time",
  ylab = "Value",
  shape_points = 20,
  size_points = 2.5,
  theme = ggplot2::theme_linedraw()
)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}})
object containing NAs. This is the ONLY necessary parameter, all other are are
only needed if you want to adjust the resulting plot.}

\item{x_axis_labels}{For adding specific x-axis labels. Takes a vector (with the same length as x)
of either Date or POSIXct objects as an input. Default (NULL) is using the
observation number as  x-axis tick labels.}

\item{color_points}{Point color for observations}

\item{color_lines}{Line color}

\item{color_missing}{Color used for highlighting the time spans with NA values}

\item{color_missing_border}{Color used as border for time spans with NA values}

\item{alpha_missing}{Transparency value used for highlighting the time spans with NA values}

\item{title}{Title of the plot (NULL for deactivating title)}

\item{subtitle}{Subtitle of the plot (NULL for deactivating subtitle)}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{shape_points}{Plotting 'character', i.e., symbol to use}

\item{size_points}{Character (or symbol) expansion: a numerical vector}

\item{theme}{Set a theme for ggplot2. Default is \code{\link[ggplot2]{theme_linedraw}}}
}
\description{
Visualize the distribution of missing values within a time series.
}
\details{
This function visualizes the distribution of missing values within a time series.
If a value is NA, the background is colored differently. This gives a good overview
of where most missing values occur.
}
\examples{
# Example 1: Visualize the missing values in x
# x <- ts(c(1:11, 4:9,NA,NA,NA,11:15,7:15,15:6,NA,NA,2:5,3:7))
# ggplot_na_distribution(x)

# Example 2: Visualize the missing values in tsAirgap time series
ggplot_na_distribution(tsAirgap)

# Example 3: Same as example 1, just written with pipe operator
x <- ts(c(1:11, 4:9, NA, NA, NA, 11:15, 7:15, 15:6, NA, NA, 2:5, 3:7))
x \%>\% ggplot_na_distribution()
}
\seealso{
\code{\link[imputeTS]{ggplot_na_intervals}},
\code{\link[imputeTS]{ggplot_na_gapsize}}, \code{\link[imputeTS]{ggplot_na_imputations}}
}
\author{
Steffen Moritz
}
