% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{na.ma}
\alias{na.ma}
\title{Deprecated use \code{\link[imputeTS]{na_ma}} instead.}
\usage{
na.ma(x, k = 4, weighting = "exponential", maxgap = Inf, ...)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}})
object in which missing values shall be replaced}

\item{k}{integer width of the moving average window. Expands to both sides
of the center element e.g. k=2 means 4 observations (2 left, 2 right) are
taken into account. If all observations in the current window are NA, the
window size is automatically increased until there are at least 2 non-NA
values present.}

\item{weighting}{Weighting to be used. Accepts the following input:
\itemize{
\item{"simple" - Simple Moving Average (SMA)}
\item{"linear" - Linear Weighted Moving Average (LWMA)}
\item{"exponential" - Exponential Weighted Moving Average (EWMA)}
}}

\item{maxgap}{Maximum number of successive NAs to still perform imputation on.
Default setting is to replace all NAs without restrictions. With this
option set, consecutive NAs runs, that are longer than 'maxgap' will
be left NA. This option mostly makes sense if you want to
treat long runs of NA afterwards separately.}
}
\description{
na.ma is replaced by \code{\link[imputeTS]{na_ma}}.
The functionality stays the same. The new name better fits modern R code
style guidelines (which prefer _ over . in function names).
}
\keyword{internal}
