% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{na.mean}
\alias{na.mean}
\title{Deprecated use \code{\link[imputeTS]{na_mean}} instead.}
\usage{
na.mean(x, option = "mean", maxgap = Inf, ...)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}})
object in which missing values shall be replaced}

\item{option}{Algorithm to be used. Accepts the following input:
\itemize{
\item{"mean" - take the mean for imputation (default)}
\item{"median" - take the median for imputation}
\item{"mode" - take the mode for imputation}
\item{"harmonic" - take the harmonic mean}
\item{"geometric" - take the geometric mean}
}}

\item{maxgap}{Maximum number of successive NAs to still perform imputation on.
Default setting is to replace all NAs without restrictions. With this
option set, consecutive NAs runs, that are longer than 'maxgap' will
be left NA. This option mostly makes sense if you want to
treat long runs of NA afterwards separately.}
}
\description{
na.mean is replaced by \code{\link[imputeTS]{na_mean}}.
The functionality stays the same. The new name better fits modern R code
style guidelines (which prefer _ over . in function names).
}
\keyword{internal}
