% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{na.random}
\alias{na.random}
\title{Deprecated use \code{\link[imputeTS]{na_random}} instead.}
\usage{
na.random(x, lower_bound = NULL, upper_bound = NULL, maxgap = Inf, ...)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}})
object in which missing values shall be replaced}

\item{lower_bound}{Lower bound for the random samples.
If nothing or NULL is set min(x) will be used.}

\item{upper_bound}{Upper bound for the random samples.
If nothing or NULL is set man(x) will be used.}

\item{maxgap}{Maximum number of successive NAs to still perform imputation on.
Default setting is to replace all NAs without restrictions. With this
option set, consecutive NAs runs, that are longer than 'maxgap' will
be left NA. This option mostly makes sense if you want to
treat long runs of NA afterwards separately.}
}
\description{
na.random is replaced by \code{\link[imputeTS]{na_random}}.
The functionality stays the same. The new name better fits modern R code
style guidelines (which prefer _ over . in function names).
}
\keyword{internal}
