% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{na.remove}
\alias{na.remove}
\title{Deprecated use \code{\link[imputeTS]{na_remove}} instead.}
\usage{
na.remove(x, ...)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}})
object in which missing values shall be replaced}
}
\description{
na.remove is replaced by \code{\link[imputeTS]{na_remove}}.
The functionality stays the same. The new name better fits modern R code
style guidelines (which prefer _ over . in function names).
}
\keyword{internal}
