% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{na.seasplit}
\alias{na.seasplit}
\title{Deprecated use \code{\link[imputeTS]{na_seasplit}} instead.}
\usage{
na.seasplit(
  x,
  algorithm = "interpolation",
  find_frequency = FALSE,
  maxgap = Inf,
  ...
)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}})
object in which missing values shall be replaced}

\item{algorithm}{Algorithm to be used after splits.
Accepts the following input:
\itemize{
\item{"interpolation" - Imputation by Interpolation}
\item{"locf" - Imputation by Last Observation Carried Forward}
\item{"mean" - Imputation by Mean Value}
\item{"random" - Imputation by Random Sample}
\item{"kalman" - Imputation by Kalman Smoothing and State Space Models}
\item{"ma" - Imputation by Weighted Moving Average}
}}

\item{find_frequency}{If TRUE the algorithm will try to estimate the frequency
of the time-series automatically.}

\item{maxgap}{Maximum number of successive NAs to still perform imputation on.
Default setting is to replace all NAs without restrictions. With this
option set, consecutive NAs runs, that are longer than 'maxgap' will
be left NA. This option mostly makes sense if you want to
treat long runs of NA afterwards separately.}

\item{...}{Additional parameters for these algorithms that can be
passed through. Look at \code{\link[imputeTS]{na_interpolation}},
\code{\link[imputeTS]{na_locf}}, \code{\link[imputeTS]{na_random}},
\code{\link[imputeTS]{na_mean}} for parameter options.}
}
\description{
na.seasplit is replaced by \code{\link[imputeTS]{na_seasplit}}.
The functionality stays the same. The new name better fits modern R code
style guidelines (which prefer _ over . in function names).
}
\keyword{internal}
