% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_deprecated_plotNA.distribution.R
\name{plotNA.distribution}
\alias{plotNA.distribution}
\title{Deprecated use \code{\link[imputeTS]{plotNA_distribution}} instead.}
\usage{
plotNA.distribution(
  x,
  colPoints = "steelblue",
  colBackgroundMV = "indianred2",
  main = "Distribution of NAs",
  xlab = "Time",
  ylab = "Value",
  pch = 20,
  cexPoints = 0.8,
  col = "black",
  ...
)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}}) object containing NAs}

\item{colPoints}{Color of the points for each observation}

\item{colBackgroundMV}{Color for the background for the NA sequences}

\item{main}{Main label for the plot}

\item{xlab}{Label for x axis of the plot}

\item{ylab}{Label for y axis of plot}

\item{pch}{Plotting 'character', i.e., symbol to use.}

\item{cexPoints}{character (or symbol) expansion: a numerical vector.}

\item{col}{Color for the lines.}

\item{...}{Additional graphical parameters that can be passed through to plot}
}
\description{
Visualize the distribution of missing values within a time series.
plotNA.distribution is replaced by \code{\link[imputeTS]{plotNA_distribution}}.
The new function is based on ggplot2 and looks better.
For a transition period of some years the old plotNA.distribution function
will still work. 
Please adjust your code accordingly to be prepared for the future.
}
\details{
This function visualizes the distribution of missing values within a time series. Therefore, 
the time series is plotted and whenever a value is NA the background is colored differently.
This gives a nice overview, where in the time series most of the missing values occur.
}
\examples{
#Example 1: Visualize the missing values in x
x <- ts(c(1:11, 4:9,NA,NA,NA,11:15,7:15,15:6,NA,NA,2:5,3:7))
plotNA.distribution(x)

#Example 2: Visualize the missing values in tsAirgap time series
plotNA.distribution(tsAirgap)

#Example 3: Same as example 1, just written with pipe operator
x <- ts(c(1:11, 4:9,NA,NA,NA,11:15,7:15,15:6,NA,NA,2:5,3:7))
x \%>\% plotNA.distribution


}
\seealso{
\code{\link[imputeTS]{plotNA.distributionBar}},
 \code{\link[imputeTS]{plotNA.gapsize}}, \code{\link[imputeTS]{plotNA.imputations}}
}
\author{
Steffen Moritz
}
\keyword{internal}
