% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNA_distribution.R
\name{plotNA.distribution}
\alias{plotNA.distribution}
\title{Deprecated use \code{\link[imputeTS]{plotNA_distribution}} instead.}
\usage{
plotNA.distribution(x, colPoints = "steelblue",
  colBackgroundMV = "indianred2", main = "Distribution of NAs",
  xlab = "Time", ylab = "Value", pch = 20, cexPoints = 2.5,
  col = "black", theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{x}{Numeric Vector (\code{\link{vector}}) or Time Series (\code{\link{ts}}) object containing NAs}

\item{colPoints}{Point color for observations}

\item{colBackgroundMV}{Background color for NA sequences}

\item{main}{Main title}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{pch}{Plotting 'character', i.e., symbol to use}

\item{cexPoints}{Character (or symbol) expansion: a numerical vector}

\item{col}{Line color}

\item{theme}{Set a theme for ggplot2. Default is \code{\link[ggplot2]{theme_minimal}}}

\item{...}{These parameters are passed to \code{\link[ggplot2]{geom_line}}}
}
\description{
plotNA.distribution is replaced by \code{\link[imputeTS]{plotNA_distribution}}.
The functionality stays the same. The new name better fits modern R code
style guidelines (which prefer _ over . in function names).
}
\keyword{internal}
