% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{plotNA.distribution}
\alias{plotNA.distribution}
\title{Discontinued - Use \code{\link[imputeTS]{ggplot_na_distribution}} instead.}
\usage{
plotNA.distribution(x, ...)
}
\description{
plotNA.distribution was replaced by \code{\link[imputeTS]{ggplot_na_distribution}}.
The new plotting function provides an improved version of the old plot, e.g. it looks better now and is better adjustable,
because it is based on ggplot2. If you absolutely want to use the old function,
you need to download an older package version. Versions 3.0 and below still have the old functions.
}
\keyword{internal}
