% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNA_imputations.R
\name{plotNA.imputations}
\alias{plotNA.imputations}
\title{Deprecated use \code{\link[imputeTS]{plotNA_imputations}} instead.}
\usage{
plotNA.imputations(x.withNA, x.withImputations, x.withTruth = NULL,
  legend = TRUE, main = "Visualization Imputed Values",
  xlab = "Time", ylab = "Value", colWithTruth = "green3",
  colLines = "black", colWithImputations = "indianred2",
  colWithNA = "steelblue2", pch = 20, cex = 2,
  theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{x.withNA}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs before imputation}

\item{x.withImputations}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs replaced by imputed values}

\item{x.withTruth}{Numeric Vector or Time Series (\code{\link{ts}}) object with the real values. (can be set to NULL if not known)}

\item{legend}{If TRUE a legend is added at the bottom}

\item{main}{Main title}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{colWithTruth}{Color of real values (truth) for the NA values}

\item{colLines}{Color of lines connecting non-NA observations}

\item{colWithImputations}{Color of imputed values}

\item{colWithNA}{Color of non-NA observations}

\item{pch}{Either an integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{cex}{A numerical value giving the size of points.}

\item{theme}{Set a theme for ggplot2. Default is \code{\link[ggplot2]{theme_minimal}}}

\item{...}{Additional graphical parameters can be passed to \code{\link[ggplot2]{geom_line}}}
}
\description{
plotNA.imputations is replaced by \code{\link[imputeTS]{plotNA_imputations}}.
The functionality stays the same. The new name better fits modern R code
style guidelines (which prefer _ over . in function names).
}
\keyword{internal}
