% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNA_imputations.R
\name{plotNA_imputations}
\alias{plotNA_imputations}
\title{Visualize Imputed Values}
\usage{
plotNA_imputations(x.withNA, x.withImputations, x.withTruth = NULL,
  legend = TRUE, main = "Visualization Imputed Values",
  xlab = "Time", ylab = "Value", colWithTruth = "green3",
  colLines = "black", colWithImputations = "indianred2",
  colWithNA = "steelblue2", pch = 20, cex = 2,
  theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{x.withNA}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs before imputation}

\item{x.withImputations}{Numeric Vector or Time Series (\code{\link{ts}}) object with NAs replaced by imputed values}

\item{x.withTruth}{Numeric Vector or Time Series (\code{\link{ts}}) object with the real values. (can be set to NULL if not known)}

\item{legend}{If TRUE a legend is added at the bottom}

\item{main}{Main title}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{colWithTruth}{Color of real values (truth) for the NA values}

\item{colLines}{Color of lines connecting non-NA observations}

\item{colWithImputations}{Color of imputed values}

\item{colWithNA}{Color of non-NA observations}

\item{pch}{Either an integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{cex}{A numerical value giving the size of points.}

\item{theme}{Set a theme for ggplot2. Default is \code{\link[ggplot2]{theme_minimal}}}

\item{...}{Additional graphical parameters can be passed to \code{\link[ggplot2]{geom_line}}}
}
\description{
Visualize the imputed values in a time series.
}
\details{
This plot can be used, to visualize imputed values for a time series.
Imputed values (filled NA gaps) are shown in a different color than the other values.
If real values (truth) for the NA gaps are known, they are added in a different color.
}
\examples{
# Example 1: Visualize the values that were imputed by na.mean in the time series
impMean.Airgap <- na.mean(tsAirgap)
plotNA_imputations(tsAirgap, impMean.Airgap)


# Example 2: Visualize the imputed values by na_locf and the true values in the time series
impLOCF.Airgap <- na.locf(tsAirgap)
plotNA_imputations(tsAirgap, impLOCF.Airgap, tsAirgapComplete)

# Example 3: Same as example 1, just written with pipe operator
tsAirgap \%>\%
  na.mean() \%>\%
  plotNA_imputations(x.withNA = tsAirgap)
}
\seealso{
\code{\link[imputeTS]{plotNA_distribution}},\code{\link[imputeTS]{plotNA_distributionBar}},
 \code{\link[imputeTS]{plotNA_gapsize}}
}
\author{
Steffen Moritz
}
